+++ RiskOfChaos/ModifierController/UI/IUIModificationProvider.cs
@@ -1,0 +1,6 @@
+﻿namespace RiskOfChaos.ModifierController.UI
+{
+    public interface IUIModificationProvider : IValueModificationProvider<UIModificationData>
+    {
+    }
+}