+++ RiskOfChaos/Utilities/CameraEffects/CameraEffectType.cs
@@ -1,0 +1,8 @@
+﻿namespace RiskOfChaos.Utilities.CameraEffects
+{
+    public enum CameraEffectType : byte
+    {
+        World,
+        UIAndWorld
+    }
+}